requirements.txt
azure-identity
aiohttp
azure-ai-projects==1.0.0b12 